Broke out useSubmit to submit file
